Prepare for merge and address PR comments
- rearrange base traits
- replace "IdiomaticMockito2" with "IdiomaticMockito.WithExpect" + "experimental" comments
- extract non-verification tests from expect/verify tests (but more can be done)
- bump version and update readme
