+++ core/src/main/scala/org/mockito/IdiomaticMockito.scala
@@ -5,18 +5,16 @@ trait IdiomaticMockito extends IdiomaticStubbing with PostfixVerifications {
   override def verification(v: => Any): Verification = v
 }
 
-/**
- * Simple object to allow the usage of the trait without mixing it in
- */
-object IdiomaticMockito extends IdiomaticMockito
+object IdiomaticMockito extends IdiomaticMockito {
 
-// TODO need a better name
-trait IdiomaticMockito2 extends IdiomaticStubbing with PrefixExpectations {
-  override type Verification = Unit
-  override def verification(v: => Any): Verification = v
+  /**
+   * EXPERIMENTAL base trait using new prefix DSL for verifications.
+   * Use with care: API may change between minor versions.
+   */
+  trait WithExpect extends IdiomaticStubbing with PrefixExpectations {
+    override type Verification = Unit
+    override def verification(v: => Any): Verification = v
+  }
+
+  object WithExpect extends WithExpect
 }
-
-/**
- * Simple object to allow the usage of the trait without mixing it in
- */
-object IdiomaticMockito2 extends IdiomaticMockito2
