+++ version.properties
@@ -1,4 +1,4 @@
 #Version of the produced binaries. This file is intended to be checked-in.
 #It will be automatically bumped by release automation.
-version=1.12.1
+version=1.13.0
 previousVersion=1.12.0
